requirements.txt
flask
pypdf
python-docx
werkzeug
